Move assigned_users and assigned_teams out from the ogds-service into the rest-service
